DELIA-59816 : RDKPerf waiting on a thread to exit
Reason for change: RDKPerf is waiting on a thread join
when an attached process exits.  This change allows the process
detach to wake up the thread and get it to exit faster.

Test Procedure: run gst-inspect and observe that there are no delays
There should also be the following log message.

[RDKPerf P 6D85 : T B5CB3460] : Task(155) : Exit task loop has been signaled

Risks: None

Signed-off-by: Douglas Adler <douglas_adler2@comcast.com>
Change-Id: I426b281df4475cedd94ae5bda487416f31083334