DELIA-61975: Update init data filtering for youtube use case
Reason for change: Current patch does not provide the init data
to the application if the init data does not contain the active
decryption system. However, Youtube content requires that init
data is still passed to the application in such case, so we need
the patch adjusted while app partners that required this patch
in the first place fix their app. This patch is intended to be
eventually removed.
Test Procedure: Verify assets play normally, after seek and
after ads
Risks: Medium

Signed-off-by: Filipe Norte <filipe.norte@sky.uk>
Change-Id: I864dc401a3798ba1e6eaa800811d316c55b0900c