RDKBDEV-3039: Fix "invalid escape sequence" warning
Python tried to interpret "\S" in strings as escape sequence.  This is
not a valid escape sequence for python strings, so was processed
literally - without any substitution, but a warning was issued. This
change retains the no-substitution behaviour, but prevents python from
scanning for escape sequences by using a raw string literal.

Reason for change: fix warning.
Test procedure: remove the "tmp" directory and run bitbake command again
Risk: None

Change-Id: I4adcd07b39aa9d5d5f8015f9b7915cb41a99e582
Signed-off-by: Dominik Wójt <dominik.wojt@cognizant.com>