SERXIONE-7573 : Test-Tubi crash in CachedResourceStreamingClient::responseRec
Reason for change: based on stack trace, it looks like the WebkitWebSrc pointer
may be pointing to garbage and results in a segmentation fault.
There was a similar issue reported and fixed in webkit upstream -
https://bugs.webkit.org/show_bug.cgi?id=266708 https://bugs.webkit.org/show_bug.cgi?id=266973 ,
https://github.com/WebKit/WebKit/commit/6a9d7a164705ebaf78a158be71f97a5f520f9655.
These changes are available in wpe-2.46 but not backported to the current webkit version of wpe-2.38.
Test Procedure: Test for regressions as issue is not reproducible.
Risks: Low

Change-Id: I0a9c6e56f82c4bdcfd86aea0524b798877726bb1
Signed-off-by: Simi Mathew <simim@tataelxsi.co.in>