RDKCOM-5371: RDKBDEV-3220 : Disabling Alternatives when syslog-ng DISTRO_FEA...
Reason for change: It is observed that there is no check for syslog-ng DISTRO_FEATURE in syslog-ng Yocto recipe.
So here, we will be introducing a check for DISTRO_FEATURE such that when feature is not enabled, alternatives should not be set in recipe.
Risks: Low
Signed-off-by: Aiswarya Prasad <aprasad@maxlinear.com>
Test Procedure: None.
Priority: P1

(cherry picked from commit 1d661bb46598063121805006855ee25c7e25e82b)
Change-Id: I988492f713e3e4936ef8f605a66336ddd7810779